+++ packages/components/src/components.d.ts
@@ -46,6 +46,39 @@ export namespace Components {
          */
         "toggle": (open?: boolean) => Promise<boolean>;
     }
+    /**
+     * @class PostIcon - representing a stencil component
+     */
+    interface PostIcon {
+        /**
+          * The name of the animation (`cylon`, `cylon-vertical`, `spin`, `spin-reverse`, `fade`, `throb`).
+         */
+        "animation"?: string;
+        /**
+          * The base path, where the icons are located (must be a public url).
+         */
+        "base"?: string;
+        /**
+          * When set to `true`, the icon will be flipped horizontally.
+         */
+        "flipH"?: boolean;
+        /**
+          * When set to `true`, the icon will be flipped vertically.
+         */
+        "flipV"?: boolean;
+        /**
+          * The name/id of the icon (e.g. 1000, 1001, ...).
+         */
+        "name": string;
+        /**
+          * The `number` of degree for the css `rotate` transformation.
+         */
+        "rotate"?: number;
+        /**
+          * The `number` for the css `scale` transformation.
+         */
+        "scale"?: number;
+    }
 }
 declare global {
     interface HTMLPostAlertElement extends Components.PostAlert, HTMLStencilElement {
@@ -60,9 +93,19 @@ declare global {
         prototype: HTMLPostCollapsibleElement;
         new (): HTMLPostCollapsibleElement;
     };
+    /**
+     * @class PostIcon - representing a stencil component
+     */
+    interface HTMLPostIconElement extends Components.PostIcon, HTMLStencilElement {
+    }
+    var HTMLPostIconElement: {
+        prototype: HTMLPostIconElement;
+        new (): HTMLPostIconElement;
+    };
     interface HTMLElementTagNameMap {
         "post-alert": HTMLPostAlertElement;
         "post-collapsible": HTMLPostCollapsibleElement;
+        "post-icon": HTMLPostIconElement;
     }
 }
 declare namespace LocalJSX {
@@ -98,9 +141,43 @@ declare namespace LocalJSX {
          */
         "headingLevel"?: number;
     }
+    /**
+     * @class PostIcon - representing a stencil component
+     */
+    interface PostIcon {
+        /**
+          * The name of the animation (`cylon`, `cylon-vertical`, `spin`, `spin-reverse`, `fade`, `throb`).
+         */
+        "animation"?: string;
+        /**
+          * The base path, where the icons are located (must be a public url).
+         */
+        "base"?: string;
+        /**
+          * When set to `true`, the icon will be flipped horizontally.
+         */
+        "flipH"?: boolean;
+        /**
+          * When set to `true`, the icon will be flipped vertically.
+         */
+        "flipV"?: boolean;
+        /**
+          * The name/id of the icon (e.g. 1000, 1001, ...).
+         */
+        "name"?: string;
+        /**
+          * The `number` of degree for the css `rotate` transformation.
+         */
+        "rotate"?: number;
+        /**
+          * The `number` for the css `scale` transformation.
+         */
+        "scale"?: number;
+    }
     interface IntrinsicElements {
         "post-alert": PostAlert;
         "post-collapsible": PostCollapsible;
+        "post-icon": PostIcon;
     }
 }
 export { LocalJSX as JSX };
@@ -109,6 +186,10 @@ declare module "@stencil/core" {
         interface IntrinsicElements {
             "post-alert": LocalJSX.PostAlert & JSXBase.HTMLAttributes<HTMLPostAlertElement>;
             "post-collapsible": LocalJSX.PostCollapsible & JSXBase.HTMLAttributes<HTMLPostCollapsibleElement>;
+            /**
+             * @class PostIcon - representing a stencil component
+             */
+            "post-icon": LocalJSX.PostIcon & JSXBase.HTMLAttributes<HTMLPostIconElement>;
         }
     }
 }