+++ packages/components/src/utils/tests/property-checkers.spec.ts
@@ -1,8 +1,15 @@
 /*
  * Copyright 2022 by Swiss Post, Information Technology
  */
-
-import { checkBoolean, checkNonEmptyString, checkOneOf, checkPattern } from '../property-checkers';
+import {
+  checkType,
+  checkBoolean,
+  checkEmptyOrType,
+  checkNonEmptyString,
+  checkOneOf,
+  checkPattern,
+  checkEmptyOrOneOf,
+} from '../property-checkers';
 
 describe('property-checkers', () => {
   let errorMessage: string;
@@ -15,6 +22,210 @@ describe('property-checkers', () => {
 
   beforeEach(() => {
     checkerParameters = [];
+  });
+
+  describe('typeChecker', () => {
+    beforeEach(() => {
+      checker = checkType;
+    });
+
+    it('should not throw an error if the value is boolean', () => {
+      checkerParameters = ['boolean'];
+      errorMessage = 'Is not boolean.';
+      [true, false].forEach(v => {
+        expect(runCheckerWithValue(v)).not.toThrow();
+      });
+    });
+
+    it('should not throw an error if the value is number', () => {
+      checkerParameters = ['number'];
+      errorMessage = 'Is not number.';
+      [42, 4.2, 4_200, 2.4434634e9, NaN].forEach(v => {
+        expect(runCheckerWithValue(v)).not.toThrow();
+      });
+    });
+
+    it('should not throw an error if the value is string', () => {
+      checkerParameters = ['string'];
+      errorMessage = 'Is not string.';
+      ['', 'string', '42', '¡¡Olé 🙌!!'].forEach(v => {
+        expect(runCheckerWithValue(v)).not.toThrow();
+      });
+    });
+
+    it('should not throw an error if the value is array', () => {
+      checkerParameters = ['array'];
+      errorMessage = 'Is not array.';
+      [[]].forEach(v => {
+        expect(runCheckerWithValue(v)).not.toThrow();
+      });
+    });
+
+    it('should not throw an error if the value is object', () => {
+      checkerParameters = ['object'];
+      errorMessage = 'Is not object.';
+      [null, {}].forEach(v => {
+        expect(runCheckerWithValue(v)).not.toThrow();
+      });
+    });
+
+    it('should not throw an error if the value is function', () => {
+      checkerParameters = ['function'];
+      errorMessage = 'Is not function.';
+      [function () {}, () => {}].forEach(v => {
+        expect(runCheckerWithValue(v)).not.toThrow();
+      });
+    });
+
+    it('should throw an error if the value is not boolean', () => {
+      checkerParameters = ['boolean'];
+      errorMessage = 'Is boolean.';
+      [undefined, null, 42, NaN, 'string', [], {}, () => {}].forEach(v => {
+        expect(runCheckerWithValue(v)).toThrow();
+      });
+    });
+
+    it('should throw an error if the value is not number', () => {
+      checkerParameters = ['number'];
+      errorMessage = 'Is number.';
+      [undefined, null, true, 'string', [], {}, () => {}].forEach(v => {
+        expect(runCheckerWithValue(v)).toThrow();
+      });
+    });
+
+    it('should throw an error if the value is not string', () => {
+      checkerParameters = ['string'];
+      errorMessage = 'Is string.';
+      [undefined, null, true, 42, NaN, [], {}, () => {}].forEach(v => {
+        expect(runCheckerWithValue(v)).toThrow();
+      });
+    });
+
+    it('should throw an error if the value is not array', () => {
+      checkerParameters = ['array'];
+      errorMessage = 'Is array.';
+      [undefined, null, true, 42, NaN, 'string', {}, () => {}].forEach(v => {
+        expect(runCheckerWithValue(v)).toThrow();
+      });
+    });
+
+    it('should throw an error if the value is not object', () => {
+      checkerParameters = ['object'];
+      errorMessage = 'Is object.';
+      [undefined, true, 42, NaN, 'string', () => {}].forEach(v => {
+        expect(runCheckerWithValue(v)).toThrow();
+      });
+    });
+
+    it('should throw an error if the value is not function', () => {
+      checkerParameters = ['function'];
+      errorMessage = 'Is function.';
+      [undefined, null, true, 42, NaN, 'string', [], {}].forEach(v => {
+        expect(runCheckerWithValue(v)).toThrow();
+      });
+    });
+  });
+
+  describe('emptyOrTypeChecker', () => {
+    beforeEach(() => {
+      checker = checkEmptyOrType;
+    });
+
+    it('should not throw an error if the value is empty or boolean', () => {
+      checkerParameters = ['boolean'];
+      errorMessage = 'Is not empty or boolean.';
+      [undefined, null, '', true, false].forEach(v => {
+        expect(runCheckerWithValue(v)).not.toThrow();
+      });
+    });
+
+    it('should not throw an error if the value is empty or number', () => {
+      checkerParameters = ['number'];
+      errorMessage = 'Is not empty or number.';
+      [undefined, null, '', 42, 4.2, 4_200, 2.4434634e9, NaN].forEach(v => {
+        expect(runCheckerWithValue(v)).not.toThrow();
+      });
+    });
+
+    it('should not throw an error if the value is empty or string', () => {
+      checkerParameters = ['string'];
+      errorMessage = 'Is not empty or string.';
+      [undefined, null, '', 'string', '42', '¡¡Olé 🙌!!'].forEach(v => {
+        expect(runCheckerWithValue(v)).not.toThrow();
+      });
+    });
+
+    it('should not throw an error if the value is empty or array', () => {
+      checkerParameters = ['array'];
+      errorMessage = 'Is not empty or array.';
+      [undefined, null, '', []].forEach(v => {
+        expect(runCheckerWithValue(v)).not.toThrow();
+      });
+    });
+
+    it('should not throw an error if the value is empty or object', () => {
+      checkerParameters = ['object'];
+      errorMessage = 'Is not empty or object.';
+      [undefined, '', null, {}].forEach(v => {
+        expect(runCheckerWithValue(v)).not.toThrow();
+      });
+    });
+
+    it('should not throw an error if the value is empty or function', () => {
+      checkerParameters = ['function'];
+      errorMessage = 'Is not empty or function.';
+      [undefined, null, '', function () {}, () => {}].forEach(v => {
+        expect(runCheckerWithValue(v)).not.toThrow();
+      });
+    });
+
+    it('should throw an error if the value is not empty or boolean', () => {
+      checkerParameters = ['boolean'];
+      errorMessage = 'Is empty or boolean.';
+      [42, NaN, 'string', [], {}, () => {}].forEach(v => {
+        expect(runCheckerWithValue(v)).toThrow();
+      });
+    });
+
+    it('should throw an error if the value is not empty or number', () => {
+      checkerParameters = ['number'];
+      errorMessage = 'Is number.';
+      [true, 'string', [], {}, () => {}].forEach(v => {
+        expect(runCheckerWithValue(v)).toThrow();
+      });
+    });
+
+    it('should throw an error if the value is not empty or string', () => {
+      checkerParameters = ['string'];
+      errorMessage = 'Is empty or string.';
+      [true, 42, NaN, [], {}, () => {}].forEach(v => {
+        expect(runCheckerWithValue(v)).toThrow();
+      });
+    });
+
+    it('should throw an error if the value is not empty or array', () => {
+      checkerParameters = ['array'];
+      errorMessage = 'Is empty or array.';
+      [true, 42, NaN, 'string', {}, () => {}].forEach(v => {
+        expect(runCheckerWithValue(v)).toThrow();
+      });
+    });
+
+    it('should throw an error if the value is not empty or object', () => {
+      checkerParameters = ['object'];
+      errorMessage = 'Is empty or object.';
+      [true, 42, NaN, 'string', () => {}].forEach(v => {
+        expect(runCheckerWithValue(v)).toThrow();
+      });
+    });
+
+    it('should throw an error if the value is not empty or function', () => {
+      checkerParameters = ['function'];
+      errorMessage = 'Is empty or function.';
+      [true, 42, NaN, 'string', [], {}].forEach(v => {
+        expect(runCheckerWithValue(v)).toThrow();
+      });
+    });
   });
 
   describe('oneOfChecker', () => {
@@ -33,6 +244,26 @@ describe('property-checkers', () => {
     });
   });
 
+  describe('emptyOrOneOfChecker', () => {
+    beforeEach(() => {
+      checker = checkEmptyOrOneOf;
+      checkerParameters = [['A', 'B', 'C', 'D']];
+      errorMessage = 'Is not empty or one of.';
+    });
+
+    it('should not throw an error if the value is empty or one of the possible values', () => {
+      [undefined, null, '', 'A'].forEach(v => {
+        expect(runCheckerWithValue(v)).not.toThrow();
+      });
+    });
+
+    it('should throw the provided error if the value is not one of the possible values', () => {
+      [true, 42, NaN, 'E', [], {}, () => {}].forEach(v => {
+        expect(runCheckerWithValue(v)).toThrow();
+      });
+    });
+  });
+
   describe('booleanChecker', () => {
     beforeEach(() => {
       checker = checkBoolean;
@@ -40,13 +271,13 @@ describe('property-checkers', () => {
     });
 
     it('should not throw an error if the value is a boolean', () => {
-      [true, false].forEach((boolean) => {
+      [true, false].forEach(boolean => {
         expect(runCheckerWithValue(boolean)).not.toThrow();
       });
     });
 
     it('should throw the provided error if the value is not a boolean', () => {
-      [undefined, null, NaN, 1, 'a', {}, [], () => {}].forEach((notBoolean) => {
+      [undefined, null, NaN, 1, 'a', {}, [], () => {}].forEach(notBoolean => {
         expect(runCheckerWithValue(notBoolean)).toThrow(errorMessage);
       });
     });
@@ -63,13 +294,13 @@ describe('property-checkers', () => {
     });
 
     it('should throw the provided error if the value is not a string', () => {
-      [undefined, null, NaN, 1, true, {}, [], () => {}].forEach((notString) => {
+      [undefined, null, NaN, 1, true, {}, [], () => {}].forEach(notString => {
         expect(runCheckerWithValue(notString)).toThrow(errorMessage);
       });
     });
 
     it('should throw the provided error if the value is an empty string', () => {
-      ['', '   '].forEach((emptyString) => {
+      ['', '   '].forEach(emptyString => {
         expect(runCheckerWithValue(emptyString)).toThrow(errorMessage);
       });
     });
@@ -87,7 +318,7 @@ describe('property-checkers', () => {
     });
 
     it('should throw the provided error if the value is not a string', () => {
-      [undefined, null, NaN, 1, true, {}, [], () => {}].forEach((notString) => {
+      [undefined, null, NaN, 1, true, {}, [], () => {}].forEach(notString => {
         expect(runCheckerWithValue(notString)).toThrow(errorMessage);
       });
     });