+++ packages/documentation/.storybook/main.js
@@ -4,11 +4,13 @@ module.exports = {
   addons: [
     '@storybook/addon-docs',
     '@storybook/addon-controls',
+    '@storybook/addon-links',
     '@storybook/addon-measure',
     '@storybook/addon-viewport',
     '@storybook/addon-a11y',
     '@pxtrn/storybook-addon-docs-stencil',
     'storybook-dark-mode',
+    '@geometricpanda/storybook-addon-badges',
     {
       name: '@storybook/preset-scss',
       options: {
@@ -18,7 +20,7 @@ module.exports = {
       },
     },
   ],
-  staticDirs: ['../public'],
+  staticDirs: ['../public', '../node_modules/@swisspost/design-system-icons/public'],
   webpackFinal: (config, options) => {
     // allow scss :export statments (scss variables to js)
     config.module.rules.forEach(rule => {