+++ packages/documentation/cypress/e2e/home.cy.ts
@@ -1,0 +1,10 @@
+describe('home', () => {
+  before(() => {
+    cy.visitStorybook();
+  });
+
+  it('should show the home page', () => {
+    cy.loadStory('Home', 'Page');
+    cy.get('.docs-header h1').should('have.text', 'Swiss PostDesign System');
+  });
+});