+++ packages/documentation/src/stories/internet-header/breadcrumbs/post-internet-breadcrumbs.stories.tsx
@@ -1,4 +1,5 @@
 import { Meta, Args, Story } from '@storybook/react';
+import { BADGE } from '@geometricpanda/storybook-addon-badges';
 import Header from '../header/post-internet-header.stories';
 import docsPage from './internet-breadcrumbs.docs.mdx';
 import '../header/internet-header.styles.scss';
@@ -10,6 +11,7 @@ export default {
     docs: {
       page: docsPage,
     },
+    badges: [BADGE.STABLE],
   },
   argTypes: {
     'custom-items': {