+++ packages/internet-header/src/services/search/search-utilities.ts
@@ -1,5 +1,3 @@
-import slugify from '@sindresorhus/slugify';
-
 /**
  * Transform to lowercase, extract accents to separate unicode chars,
  * replace all accent unicode chars so only standard latin chars remain
@@ -13,19 +11,3 @@ export const hardNormalize = (str: string): string =>
     .toLowerCase()
     .normalize('NFD')
     .replace(/[\u0300-\u036f]/g, '');
-
-/**
- * Convert a poi name to a URL compatible slug, logic copied from TOPOS
- * @param fragment A poi name
- * @returns
- */
-export const createSlug = (fragment: string): string =>
-  slugify(
-    fragment
-      .replace(/ \/ /g, '-')
-      .replace(/ /g, '-')
-      .replace(/\//g, '')
-      .replace(/\./g, '')
-      .replace(/\,/g, '')
-      .replace(/\'/g, ''),
-  );