+++ packages/components-angular/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @swisspost/design-system-components-angular
 
+## 1.0.8
+
+### Patch Changes
+
+- Updated dependencies:
+  - @swisspost/design-system-styles@5.3.2
+
+## 1.0.7
+
+### Patch Changes
+
+- Updated dependencies:
+  - @swisspost/design-system-styles@5.3.1
+
 ## 1.0.6
 
 ### Patch Changes