+++ packages/demo/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @swisspost/design-system-demo
 
+## 5.2.3
+
+### Patch Changes
+
+- Updated dependencies:
+  - @swisspost/design-system-styles@5.3.2
+  - @swisspost/design-system-intranet-header@3.0.8
+
+## 5.2.2
+
+### Patch Changes
+
+- Updated dependencies:
+  - @swisspost/design-system-intranet-header@3.0.7
+  - @swisspost/design-system-styles@5.3.1
+
 ## 5.2.1
 
 ### Patch Changes